Reorganize directory structure and minimize setup requirements.
Move Python code to lib/, to have the same structure in the source
and the installed code.

Make the init script and all bin/ scripts recognize their location
and add the nicos code to sys.path automatically.  This removes the
need to have /etc/nicos.conf and to put the root dir into nicos.conf.