[Feature] Add wiki link possibility to NicosError
The constructor accepts a the keyword 'wikicode' with an integer
argument to create a link to a wiki page
   https://trac.frm2.tum.de/projects/NICOS/wiki/NicosError/<integer>
where more information can be given.

The integer should be the unix timestamp (e.g from 'date +%%s', to
get a uniqe id) of the first use of this specific error.

To create the wiki page, log in to trac and
enter 'wiki:NicosError/<integer>' in the search box on the upper
right.

Change-Id: I531e3d9493b13d9740cc17d2438f7cbd6ce2ab83
Reviewed-on: https://trac.frm2.tum.de/review/384
Tested-by: JenkinsCodeReview <bpedersen@frm2.tum.de>
Reviewed-by: Georg Brandl <georg.brandl@frm2.tum.de>