INSTALL : Modify the setups install
The installation command for the instrument specific setups is modified.
If the 'cp' command does not support the -n option, then the installation of
setup files is done by individual checks whether the file in the sources is
newer than the in the installation path and then copies the files.

Change-Id: I74e08f17a8fefa4596ce3eec1891a747bc26b418
Reviewed-on: https://trac.frm2.tum.de/review/888
Tested-by: Jens Krueger <jens.krueger@frm2.tum.de>
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>