TOFTOF: create a switch device for the CCR boxes.
The current implementation of the CCR box (esp. the Beckhoff PLC program)
does not allow to set the switch on and then read back.

You have to use a device switch simulates a 'key press' event via software
on the PLC and the current state is available via a different taco device.

Change-Id: I17b371b6b656b5d92bfb6c1791706f42c91e0bc7
Reviewed-on: https://trac.frm2.tum.de/review/936
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Tested-by: Jens Krueger <jens.krueger@frm2.tum.de>
Reviewed-by: Georg Brandl <georg.brandl@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>