Allow exponential values for ttl
At the moment they are treated as part of the key.
This creates huge amounts of files until the server stops accepting network connections and later crashes and cant be restarted, since it opens too many files for the network sockets to get filedescriptors below the 1024 limit, so no network connections will work, until one deletes the bogus files in the cache-db-datadir.
This basically fixes a bug which is triggered by a workaround for the cache-latency-bug (namely setting the maxage of the tas-device to a very small number (1e-20 works)

Change-Id: Iee6156665dbf129dcf2888095c7361452ed49099
Reviewed-on: https://trac.frm2.tum.de/review/1220
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Bjoern Pedersen <bjoern.pedersen@frm2.tum.de>
Reviewed-by: Georg Brandl <georg.brandl@frm2.tum.de>