count: special handling of result printing object
This is an attempt to remedy two problems with a single counting using count():

* in scripts, if you do not use ``print count(...)``, the result will be lost
* the return value is a simple list of values (time/counts/...), which is great for scripts but not useful to the user

This patch wraps the result in an object that behaves like a list but has a special __display__
method that is checked by the display hooks of the session and its printout is muted.
It then adds a printinfo() call to the count function that outputs the result in a
human-readable way.

Change-Id: I6750301b4142da22d09f7833da06b6a61f743a05
Reviewed-on: https://trac.frm2.tum.de/review/1703
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Georg Brandl <georg.brandl@frm2.tum.de>