tas: add eulerian cradle code
At the moment, the code is integrated as loosely as possible:
there is a separate Euler device that allows to define a scattering
plane by moving it at a position consisting of two vectors that
should lie in the plane.  The chi and phi circles are then moved
to the calculated position, and the omega/psi circle is moved by
the TAS code as usual.

This of course does not take full advantage of the cradle, as e.g. alternate
positions during a scan (e.g. to avoid putting the cradle in the beam)
cannot be taken into account once the scattering plane is set.

Change-Id: Ifd638dbfe2271d6a381ba781e3a6c3c542ca92cc