PANDA: Implemented workaround for buggy SPS
This is a QnD Hack, but is tested to work.
We map a NC-Error to an OK-statuscode in doStatus and
rely on the generic.Axis to call setPosition which we use to
acknowledge the NC-Error and stop the SPS.
The axis then retries the positioning which (so far) usually succeeds.

side-effects:
* a reset als trys to acknowledge an error.
* every NC-Error uses up one of the maxtries.
* we needed to set jitter quite high, as the sps switches between reporting steps and real position as the readout which causes 'not moving to goal' errors all over the place after an NC-Error.
Maybe the generic.Axis should reset the lastpos value before restarting the posiioning (calling motor.start)

The other possible implementation could have acknowledged the Error from inside doStatus, but I didn't want doStatus to actually change the Status.
It should only report the status.

Further Improvements:
The Hack could be disabled by a Param

Change-Id: If244385e3670fd6bc80facc71d3eba1277de5a49
Reviewed-on: https://trac.frm2.tum.de/review/2099
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Tested-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>