NICOS: updated Cachewriter to better work with the 15T magnet
* labview is updated to also give status (before the 30s waiting to
  open the heatswitches and after this), maxage and pollinterval
  need to be above 34s for this to work properly
* CacheWriter also does not crash at doing a wait after a freshly
  initialized NICOS (was accessing self._goal which was in this case
  not initialized)
* reworked the waiting loop in doWait

Maybe it would make sense to start a value polling thread deriving a
'stable' and 'stable and within target' criteria whose are used in doStatus
whenever there is no valid status in the cache (otherwise we take that one).
The we could use the default wait which just checks the status.

Change-Id: I786d27cf4c76283282aedea1c9d34240ce30d8f6
Reviewed-on: https://trac.frm2.tum.de/review/2185
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Bjoern Pedersen <bjoern.pedersen@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>