core device: add a "warnlimits" parameter
This parameter can be set to a range of values (e.g. (0, 50) for a Beryllium
filter temperature) that is "normal" during operation.

Currently no action is taken if the value is outside.  This parameter is
supposed to supersede the "min" and "max" configuration in the status monitor,
which always felt wrong to configure there and not where the device itself is
configured.  This parameter will also be used in the pyfrid-style panel/tree of
all devices.

Whether the NICOS commands such as read() should emit a warning should be
discussed.  It can easily become a nuisance, so there would have to be a way to
acknowledge and disable the warning.

Change-Id: I0029225e8fbfed5241aa4d65de3e6e8a32ecb6fc
Reviewed-on: https://trac.frm2.tum.de/review/2263
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>