CORE: fix bug in nicos.devices.tas.mono
Upon changing the unit, make sure the new userlimits are within the
new abslimits or a ConfigurationError will be raised *after* setting
the new abslimits, but *before* the unit is really set. This is
annoying and there is no way out other than modifying values in the
cache directly.

This patch fixes #635.

tested at panda.

+fixed accidentially changed pylintrc.

Change-Id: I573963cbc179a94abd5e2f4a26479d43fc831229
Reviewed-on: https://trac.frm2.tum.de/review/2373
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Tested-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>