utils: fix pickling/unpickling of readonly{list,dict} objects
These objects are pickled when they are sent from daemon to client.
Unpickling them as readonlydicts/lists failed because in the pickle
protocol list/dict items are pickled separately and the final object
is built elementwise. In this instance the easiest thing is to
let them pickle as normal dicts/lists.

Change-Id: Ie1bdba383b05b8f8825cb84db616a9bdaa6a02a0
Reviewed-on: https://trac.frm2.tum.de/review/2441
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Reviewed-by: Georg Brandl <georg.brandl@frm2.tum.de>