device: revert https://trac.frm2.tum.de/review/#/c/2407/
MIRA has a trivial basic instrument setup in "system" with a no-op Sample.
And in the "tas" setup there is the real TASSample.

Now when the daemon is restarted, it first loads "startup" (with "system")
so Sample is of class Sample, then loads "tas" as the last master setup
so Sample is now TASSample, and all parameters of the sample are cleared.

We have to discuss if there is a solution to this.

Change-Id: I20a1a9bffbb37d5d66482e6cb078a2dd21bc0147
Reviewed-on: https://trac.frm2.tum.de/review/2463
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Georg Brandl <georg.brandl@frm2.tum.de>