Device list panel: Add option for 'no icons' display
Add the option 'icons', which is default true, to enable an icon
less display. The states will be display in the background color of
the device name. So the eye of the user will be lead more or less
directly to the problems of the instrument.

The setup noiconsdevlist.py demonstrates the view.

Change-Id: I832c563180d6aec53c09cdea59d59d2f51cb8b85
Reviewed-on: https://trac.frm2.tum.de/review/2924
Reviewed-by: Georg Brandl <georg.brandl@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>