CORE: implement a default doWait in class Moveable
and update docu.
Also improve waitForStatus to output the current Position and Status during
the waiting loop if the device.loglevel is 'debug', as request by the ticket.

Fixes: #975

Should have no side-effects on already existing devices as they have their
own doWait(), often just a version of waitForStatus.

The new behaviour is a follows:
The default implementation just uses waitForStatus to poll status
until it is no longer BUSY. If there is no doStatus, and the default doWait,
the call to doWait is skipped, as status would return status.UNKNOWN which
is treated as 'not BUSY' by waitForStatus anyway.
If there is inconsistently a custom doWait, but no doStatus, a warning is generated,
as this seems not make much sense and should be avoided. doWait is still called,
since functionality is more important.
-> no change of functionality
Some instantly switching devices have a default implementation of doStatus,
which always returns status.OK
-> no change of functionality,

After merging this, a code cleanup is suggested in:
- nicos.devices.generic.slit.SlitAxis (doWait without doStatus)
- nicos.devices.tas.spectro.TASIndex (-''-)
- all classes which basically just mimic or use waitForStatus

Big thanks to the test suite which found the SlitAxis and TASIndex inconsistency!

Tested in demo mode with virtual devices and with the testing suite.

Change-Id: I85c7328a241438a8040b156edd398583cd9854d9
Reviewed-on: https://trac.frm2.tum.de/review/3036
Reviewed-by: Alexander Lenz <alexander.lenz@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Bjoern Pedersen <bjoern.pedersen@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>