Switcher: allow configuring a default reading value
This is needed for the JCNS-Tango-servers which sometimes
yield an invalid value during movement.
With this patch, invalid values are mapped onto the value of the
default parameter to avoid needlessy raising PositionErrors upon read.

If the default parameter is not set, the behaviour is unchanged.

The functionality is needed latest at 22.7.2013.

Change-Id: I4ad9443feef1a57c44bf74482c2e00f48c7cecda
Reviewed-on: https://trac.frm2.tum.de/review/3089
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Tested-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>