GUI: Add a -h|--help option to the command line arguments.
The command line arguments for the GUI are extended by a help option.
Additionally the long options are added.

The getopt package was change because the optparse is deprecated
since python 2.7 and the argparse package was introduced in 2.6.
The getopt package exists in both.

If we agree that we can live with the deprecated package optparse
I will change this.

Change-Id: Id07205d86f905746fe445a6f7f63edec745164f0
Reviewed-on: https://trac.frm2.tum.de/review/3143
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>