Core: Improve handling+bootstrap of DeviceAlias
formerly, Alias.name could be wrong and accesses to the AliasDevice
sometimes failed, though the .alias was set correctly.
Also, sometimes the .name was wrong, resulting in a
AddEnvironment(T) to add the wrong device...

Tested with a clean as well with a dirty cache with bogus entries.

fixes: #1006

Found by pair programming with Björn Pedersen.

Change-Id: Iee81d4d6600841f9ff0f3c453fad5b346b155fbb
Reviewed-on: https://trac.frm2.tum.de/review/3252
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Reviewed-by: Georg Brandl <georg.brandl@frm2.tum.de>