Experiment: enable generation of report templates
fixes: #779

The required template file is not included as we need a mechanism for this first.

Also more rework is needed regarding obtaining proposal information and keeping
them synchronized... (will be done in another patch)

tested in demo mode with the file from:
http://www.mlz-garching.de/files/2013_template_expreport_rtf.rtf
renamed to report_{{proposal}}.rtf and put into the configured templatedir.

Change-Id: If14ff5c24c1942b7127467f0390f8c4c6de92523
Reviewed-on: https://trac.frm2.tum.de/review/3270
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>