watchdog: allow giving different setup names for starting multiple watchdog copies like for the monitor
Note that all watchdog entries in the cache share the same namespace.
That means that if multiple watchdogs are running, the messages from
multiple watchdog instances will overwrite each other in the global
"current warnings" list.

However, for the envisioned use case (using just the notification
feature, which does not use the cache entries), this does not matter.

Change-Id: I25f4ac48362a53bb792e4dbd382192d13a5886f7
Fixes: #1070
Reviewed-on: https://trac.frm2.tum.de/review/3338
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Alexander Lenz <alexander.lenz@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>