DEVICES: create a 'LockedDevice'
which, prior moving itself, moves another device to a preconfigured
unlockvalue. after movement is done, the other device will either be moved
to a preconfigured value or (if that value is None) to the previous value.

New: movement sequence is done in a thread to avoid blocking.
This may lead to interesting constellations and bugs....

implemented as mixin (which 'just' requires additional definitions of devices
whose implementation is otherwise empty.)

As this functionality is required on several instruments this goes to
lib/nicos/generic/locked.py

WARNING: There still might me some 'mixing' issues which I haven't discovered.

tested in demo mode.

Change-Id: I09ccd54919c579b66e336e71d2bf68588fd6d78d
Reviewed-on: https://trac.frm2.tum.de/review/3404
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Reviewed-by: Georg Brandl <georg.brandl@frm2.tum.de>
Reviewed-by: Bjoern Pedersen <bjoern.pedersen@frm2.tum.de>