Experiment: change managerights to a dict...
containing the requested changes upon Finish/NewExperiment.

Working backup services request avoidance of chmod'ing a file to 0.
So we need to run the daemons as a special user ('nicd').

To easy testing the integraion and to allow more flexible configuration
of what should be done when, managerights changed to None_or(dict).

Allowed keys are:
- enableDirMode, enableFileMode
- disableDirMode, disableFileMode
- owner, group

The modes are applied directly (i.e. are no masks).
owner and group can be strings (which get translated to) or numbers.

Panda needs this to be backported to 2.4 !

Change-Id: I59592fbc2c9c77e6bac64260f889cb10ace3816c
Reviewed-on: https://trac.frm2.tum.de/review/3544
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>