Devices: generic: Locked devices fixed.
Reimplemented and derived from a more generic sequencer as discussed
in tuesdays meeting.

also a new demo-device called ld supplied.

A few issues with interaction with poller and device/mixin questions still need to be adressed, but this needs another meeting.

tested in demo mode, works as expected, even when polled.

Change-Id: Ie43cc0caff6bb03cf61aad7ee645a6f971dee1e1
Reviewed-on: https://trac.frm2.tum.de/review/3552
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>