CORE: support specifying umask in nicos.conf
and support to specify this in make.conf.

This is the backport to 2.4.

umask is applied after uid/gid change in the startup.
This is a much better place than doing this later on.
Unfortunately, the umask needs to be set by the process itself....

PANDA needs this functionality backported to release-2.4!

Change-Id: I68f02751412b197dcf333ee72152481711bedd8d
Reviewed-on: https://trac.frm2.tum.de/review/3656
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>