poller session: create a subclass of CacheReader that prevents resetting parameters
Symptom: when you start the poller, the unit of a device with an alias pointing to it is set to "".

This is because the poller loads (e.g.) the alias_T setup, which creates a CacheReader, which tries
to initialize its unit parameter, finds none in the cache and uses the default "".

Change-Id: I307db5bf8022d0f83b16bd3dce7cbe549637596c
Reviewed-on: https://trac.frm2.tum.de/review/3918
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
(cherry picked from commit 0f43fb21579eaa27609083b0b153acf81e35e559)
Reviewed-on: https://trac.frm2.tum.de/review/3943