daemon ListAuthenticator: allow empty user name/passwd
entries in the password list, which accepts any username.

This will be used for MIRA: at the moment we allow any user with any
password as ADMIN.
However, I want to reduce that to GUEST privileges and have only one
ADMIN user configured.
Normal users should still be able to use any name.
If there is no password given for an account it gets at maximum USER
access.

logic is now as follows:
if user and pw fields are configured, check credentials, use configured userlevel
else if no password is configured and username matches, restrict to at most USER
else if no user is specified, but a password is set, restrict to at most USER
else if neither user nor password is set, restrict to GUEST

Change-Id: Ibbea2f3cfa600d6fe25734498795386aba99168a
Reviewed-on: https://trac.frm2.tum.de/review/3924
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Tested-by: Jens Krueger <jens.krueger@frm2.tum.de>
Reviewed-by: Georg Brandl <georg.brandl@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>