core: simplify countloop, use doSave to call saveImage in ImageProducers
This removes awareness of ImageProducers from the countloop.
The setting of "endtime" is also moved to saveImage.

Logic of the countloop is slightly altered to do the save inline for
each detector as opposed to once at the end (might save a little time).

Also in case of errors, save() is attempted nonetheless.

Change-Id: I0aaeee68a55a7e0d40781629e59e0c29314e54c0
Reviewed-on: https://trac.frm2.tum.de/review/4033
Reviewed-by: Georg Brandl <georg.brandl@frm2.tum.de>
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>