PANDA: rework workaround for buggy IPC-cards
to only be active for non-encoded axes as this made problems
on the encoded ones. Also document the bugs a little.

Should work as it is a mix of already working, or has been working before
settings.

Also inlcudes a so-far-untested maybe-workaround class.
If in need this can easily be used be renaming at the instrument (PANDA).

Change-Id: Ia07d187c20cea6abcf5a10e8bcbcedc2ba08d423
Reviewed-on: https://trac.frm2.tum.de/review/4251
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
(cherry picked from commit b0a5abecf9fbb1d6fccb15e8b9000d9b91efa207)
Reviewed-on: https://trac.frm2.tum.de/review/4253
Tested-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>