QT-Monitor: add key event for r
to refresh/redraw/fix the shown layout.
On my system the layout is always bad/broken and sometimes
causes X-Protocol errors until i press 'r' now.

automatic relayouting is still broken, this patch merely
gives you a chance to trigger this with the press of a key.

tested on my machine, should not break anything.

Change-Id: I243e3859d32a31da0af7ee94d9775a7725cf9d26
Reviewed-on: https://trac.frm2.tum.de/review/4278
Reviewed-by: Georg Brandl <georg.brandl@frm2.tum.de>
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>