core: change mechanism of finding setup files and instrument specific configuration
This is for the goal of bringing the checkout and the installed NICOS
closer together.

Until now, the instrument we're running on is detected at *install*
time.  With this change, it is detected at *runtime* (if not explicitly
configured).  Places to configure is either the local nicos.conf (the
one not in the repository, at the root of the checkout) or the process
environment.

Instead of a make.conf to be used after install only, we will have an
"instrument-specific" nicos.conf in each custom/instrument directory,
with default values that can still be overridden in the local nicos.conf
for special installations or development.

For JCNS specific installs, we allow to customize where to find the
"custom" directory; to be configured by the "custom_paths" value in
nicos.conf.

Also changed is the way setups are found: so far, the instrument's
setups are copied to ROOTDIR/setups on installation, now we will keep
them under custom/instrument, and the nicos.conf points there.

We currently copy the frm2 setups into the instrument setups; this has
to stop -- instead multiple setup paths can be given by "setup_groups"
in nicos.conf; these will point to subdirs of the custom path in use.

The init script uses the same nicos.conf files for determining which
services to run.

Fixes: #1270
Change-Id: I39192125f7a3e00e19d959d6763ddba3cd16823e
Reviewed-on: https://trac.frm2.tum.de/review/4312
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Christian Felder <c.felder@fz-juelich.de>