MultiSwitcher: support for (optional) readables
and thus make it more explicit, which devices are moved and which are only
read to obtain 'positional' information.

Needed for Sans1 background slits as they are mounted on the collimatin tables
(moving in and out with the collimation). So their position depends on yet
another device, but moving them shall not move that other device
(which is a Moveable) as well. So the implicit distinction based on
Readable vs. Moveable does no longer work correctly here.

first, untested iteration.

Change-Id: I7d1b48070aa79ebd6b4bd4c4808d22cdc9741a0d
Reviewed-on: https://trac.frm2.tum.de/review/4516
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Reviewed-by: Bjoern Pedersen <bjoern.pedersen@frm2.tum.de>