Solve menu and toolbar issues for tabbed layout
This is based on patch 2947 by Bjoern. The tabbed layout
should show the correct menus and toolbars in the top
level window depending on which tab is active. In case
a tab was detached, menus and toolbars should also move
to the newly created window.

This seems to work now, please test with

bin/nicos-demo -c custom/jcns/guiconfig.py

in the source dir and give me feedback.

Change-Id: I00735f5e9c283b54eff00a6677d3500f91e5949c
Reviewed-on: https://forge.frm2.tum.de/review/4798
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Tested-by: Jens Krueger <jens.krueger@frm2.tum.de>