Poller: reduce log spamming upon fail to start
in case of specifying a non_exisiting_setup_name in the
always_poll field, the poller for that setup crashes as it can't load
it. But it was immediately restarted.
This spammed the poller log with ~15 lines and the cache log with 2 lines,
quickly filling up any available space for logfiles.
if the logspace runs out of space, no nicos service is able to start.

This patch introduces more detailed output in such a case and then
behaves like a poller with an empy setup, i.e. it will react to setup changes
and to setup file changes.

Change-Id: I438cbd5bc260ad5d95796e20389b254247fc22a5
Reviewed-on: https://forge.frm2.tum.de/review/4891
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Tested-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>