BIODIFF: Measureable (ImagePlateDetector) doPrepare.
Imageplates needs some preparation before counting. In order to start
the counting instantly when requested the preparations should be done
in a seperate prepare step - doPrepare. For the count command it does
not matter wether this is done in doStart or in doPrepare but for scan
commands which starts other devices simultaneously it is necessary to
seperate this.

Change-Id: Iac31f4d0cf56d2c2038dcdd38bf118646fa71010
Reviewed-on: https://forge.frm2.tum.de/review/5165
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Christian Felder <c.felder@fz-juelich.de>
Tested-by: Christian Felder <c.felder@fz-juelich.de>