Measureable, count, scan: Preparation step - doPrepare.
Some Measurables needs some preparation before counting, e.g. Image
plate detectors (erasing, moving to expo position). In order to start
the counting instantly when requested the preparations should be done
in a seperate prepare step - doPrepare. For the count command it does
not matter wether this is done in doStart or in doPrepare but for scan
commands which may starts other devices simultaneously it is necessary to
seperate this.

Change-Id: Ib38ffdbde6a75ea542e69e64c883974ac762e164
Fixes: #1384
Reviewed-on: https://forge.frm2.tum.de/review/5215
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Christian Felder <c.felder@fz-juelich.de>
Tested-by: Christian Felder <c.felder@fz-juelich.de>