RScan: Disable SW based micro stepping if move and wait.
If the movement and counting is not done simultaneous do not use
software based micro stepping, e.g. moving to start point.
This is specific to nicos.biodiff.motor.MicrostepMotor. It would be
nice to move this code to nicos.core.scan.Scan to have this functionality
for all scan commands but probably this should not contain
code/functionality which is currently instrument specific.

Change-Id: I75bc86b429797dd1ee521ee59ef4114be28faa8c
Reviewed-on: https://forge.frm2.tum.de/review/5243
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Tested-by: Christian Felder <c.felder@fz-juelich.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Reviewed-by: Christian Felder <c.felder@fz-juelich.de>