cache client: fix a long standing bug in alias handling
Since rewrites handle cache keys, they are all lowercase.

This may have caused some internal inconsistencies we have seen in the past.
Basically it failed to update the local cache copy of alias devices when
putting a new value into the cache.

Because it only happens for locally generated values, it usually does not
appear for device value and status, which are mostly generated by the poller.

Change-Id: I67644424ee84fdae1b6e86567cfc69dce0315bf7
Reviewed-on: https://forge.frm2.tum.de/review/5982
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Reviewed-by: Georg Brandl <georg.brandl@frm2.tum.de>