CORE: move classes from utils.messaging to core.sessions.simulation
These classes are not general utilities, but specific to simulation mode,
they should not be in a util module.

Also move SimulationSession to the new module, since it makes sense to be there now.

nicos.utils.messaging now just contains the ZMQ context, which should be
global for the process, and may in future be filled with general ZMQ utils.

Change-Id: I51a7e7aa87cfffa9d73de2134f53ac6ce7c80f1d
Reviewed-on: https://forge.frm2.tum.de/review/5995
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Bjoern Pedersen <bjoern.pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>