[TOFTOF] Add the device to move the blades of the focusing neutron guide
The focusing neutron guide of the TOFTOF instrument consists of 4
glass blades. These blades will be mechanically bended by a piezo
driver.

There is no encoder to find out the position and the movement is not
reproduceable due to the different forces of the glass blades. Therefore
each movement references first and then goes a certain number of steps.

The calibration has to be done in a next step as well as the addition of
a sequencer mixin to allow no blocking moves.

Change-Id: I1dc3393047e31dfbbb0f9c02d585b1590e390c9f
Reviewed-on: https://forge.frm2.tum.de/review/6135
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Tested-by: Jens Krueger <jens.krueger@frm2.tum.de>