[CI] Improving general performance.
This improves both the communication and loading performance of
the cache inspector by a lot.
Testing on a sum of 39,000 entries it takes about 1.5 second to receive
the data from the server and format it as needed.
Populating the QTreeWidget with those entries takes about 0.3 second.
Because of performance reasons categories do not show the keys of their
childs anymore.

Change-Id: I059cb129a5483f29ce92ab76a75cb617f25249c2
Reviewed-on: https://forge.frm2.tum.de/review/6254
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>