CORE/SCAN: fix updating header values too early.
do this now AFTER moving to the next scanpoint.
Otherwise the nicos-headers for image files written during a scan
contain the values of the previous point, not the current one....

* typos fixed
* handle case of empty updatedicts correctly

fixes: #1648

WARNING: due to the twisted way of the imaging stuff, there are
actually TWO invocations of updateHeaderInfo().
Basically we need a clean rewrite of this. atm, it works.

WARNING2: please check that it still works with Image drums....

works @ sans1

Change-Id: Ibfe4dd009ea65564c46fd69a98b64dd296570c2e
Reviewed-on: https://forge.frm2.tum.de/review/6409
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Bjoern Pedersen <bjoern.pedersen@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>