TACO: Improve error handling in case of RPC time out
The RPC time out error occurs mostly in case of a network problem so
it would be helpful to retry the call in a such a case since the
network problem is mostly not permanent. This avoids a setting of the
tacotries parameter to values different from 1.

This could be a solution for the sometimes occuring network overload
on the instruments.

Change-Id: I8ac37588dfb3655fcfe44076b0ac216416e70049
Reviewed-on: https://forge.frm2.tum.de/review/6733
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Georg Brandl <georg.brandl@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Reviewed-by: Bjoern Pedersen <bjoern.pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
(cherry picked from commit 3b9ac36df39ff3bc5f16cf9f67ff60dc1e24c0f1)
Reviewed-on: https://forge.frm2.tum.de/review/6852