generic slit: introduce "4blades_opposite" opmode
The '4blades' mode always works the same regardless of the
coordinatesystem of the lower level devices: left/right and top/bottom
have the same sign.

To move the slit as desired on DNS, i.e. in hardware coordinates of the
slit blades with opposite coordinate system, we need a new opmode.

Change-Id: I64f0a6d5d4a8d4c2e91758fe35627c3dbba8e646
Fixes: #1738
Reviewed-on: https://forge.frm2.tum.de/review/6530
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Bjoern Pedersen <bjoern.pedersen@frm2.tum.de>
Reviewed-by: Georg Brandl <georg.brandl@frm2.tum.de>