session: fix (hopefully) lock timeouts during testing
If the cache had not locked the master lock already, the code in the session
did a lock with default timeout, which was 1800 seconds.  Of course this lock
then is never released in time.

Also set default timeout to 10 minutes instead of 30.  Should now only be
used for _cachelock_ functionality.

Change-Id: Icf7424fd9f69209c8073b9d709c282fe6b0a56cf
Fixes: #1775
Reviewed-on: https://forge.frm2.tum.de/review/6866
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Bjoern Pedersen <bjoern.pedersen@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
(cherry picked from commit b245971021744acef84f3b169d2ff7fb41a29ae3)
Reviewed-on: https://forge.frm2.tum.de/review/6880
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Reviewed-by: Georg Brandl <georg.brandl@frm2.tum.de>