daemon: fix bug with never closing event connections
If the "stop_queue" is the only object in the queue, the _qsize() will
be zero since the string in stop_queue is empty.  This leads to the
queue never yielding the "stop_queue" object!

This may well be responsible for the strange memory leak at MIRA if
lots of clients are opened and closed, maybe due to a flaky connection.

Change-Id: Ic3d6d5ed8727cd4881cfcd78ecc0825944a7a3fa
Reviewed-on: https://forge.frm2.tum.de/review/7255
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>