tools/setupfiletool: TreeView as file browser
Tool now displays all directories of */nicos-core/custom/ in a tree widget.
All directories have their setup files, wether in subdirectories or not,
displayed as child items. Those setup files have their devices as child items.

Users can select an item in the treeview and when it's either a setup file or
a device, the tool will display information about it.

Change-Id: If78692fc34cfab7bda66f528c870bcfec27ffa92
Reviewed-on: https://forge.frm2.tum.de/review/7862
Reviewed-by: Andreas Schulz <andreas.schulz@frm2.tum.de>
Tested-by: Andreas Schulz <andreas.schulz@frm2.tum.de>
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>