tools/setupfiletool: tool recognizes when data changes
When the data of a setup is changed, the tool now prepends the setup in the
tree with an asterisk to mark it (eg system.py -> *system.py).
When attempting to switch the setups, the tool asks the user wether to save
the changes to the file.
This also works with manually loaded files, but no name is updated as there is
no corresponding item. The dialog still pops up.
Also includes some bug fixes concerning manual loading of setups.
Also, when activating an item, now checks the item type, which is much safer.
Change-Id: I7088d2e0e41645add72af4dbfeb63a45fdfffe15
Reviewed-on: https://forge.frm2.tum.de/review/7956
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>