tango: add power supply
This is derived from HasTimeout, since power supplies that do not use a ramp
usually need a certain time to actually supply the desired current/voltage.

It also has the advantage that if the device does not arrive at the preset
current/voltage due to reaching limits, NICOS will realize that and raise
a positioning error.

Change-Id: I3fb7490dd7a033f412e963d8b1b40b8e7d8d8ad1
Reviewed-on: https://forge.frm2.tum.de/review/8159
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>