tools/setupfiletool: Drag and drop for devices
Users can now drag and drop devices between setup files.
Only devices can be dragged and they can only be dropped on setups.
Performing an d&d event will automatically delete the device from the
original setup and move it to the target (in the view).

Triggering the Save action then saves the changes back to the files on
the disk (as expected).

Change-Id: Icaac499b45d21a338e2aea3d9ae13326f5f8ed1b
Reviewed-on: https://forge.frm2.tum.de/review/8316
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Alexander Lenz <alexander.lenz@frm2.tum.de>
Tested-by: Alexander Lenz <alexander.lenz@frm2.tum.de>