daemon: handle unhandled script exceptions more gracefully
It is more important to keep the script thread running than to
have an error logged.

In the case we really get an unhandled exception, try to inform
the user that he needs to restart the daemon.

Change-Id: Ida6946e41c3345103cd71cabbc1a408733db32c1
Reviewed-on: https://forge.frm2.tum.de/review/8360
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Christian Felder <c.felder@fz-juelich.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>
(cherry picked from commit e6bb62207839e8e9c5620648b3063dd1bff7f057)
Reviewed-on: https://forge.frm2.tum.de/review/8382